+++ src/Navbar.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ModuleDropdown from "./ModuleDropdown";
-import logo from './media/logo_blink.svg';
+import logo from './assets/logo_blink.svg';
 
 export default function Navbar() {
     return (
