+++ src/assets/main.css
@@ -3958,11 +3958,11 @@ video {
 }
 
 #app .bg-about{
-  background-image: url("../media/aboutCurve.svg");
+  background-image: url("../assets/aboutCurve.svg");
 }
 
 #app .bg-landing-page{
-  background-image: url("../media/background_landing_page.svg");
+  background-image: url("../assets/background_landing_page.svg");
 }
 
 #app .from-transparent{
@@ -24504,11 +24504,11 @@ video {
   }
 
   #app .sm\:bg-about{
-    background-image: url("../media/aboutCurve.svg");
+    background-image: url("../assets/aboutCurve.svg");
   }
 
   #app .sm\:bg-landing-page{
-    background-image: url("../media/background_landing_page.svg");
+    background-image: url("../assets/background_landing_page.svg");
   }
 
   #app .sm\:from-transparent{
@@ -45020,11 +45020,11 @@ video {
   }
 
   #app .md\:bg-about{
-    background-image: url("../media/aboutCurve.svg");
+    background-image: url("../assets/aboutCurve.svg");
   }
 
   #app .md\:bg-landing-page{
-    background-image: url("../media/background_landing_page.svg");
+    background-image: url("../assets/background_landing_page.svg");
   }
 
   #app .md\:from-transparent{
@@ -65536,11 +65536,11 @@ video {
   }
 
   #app .lg\:bg-about{
-    background-image: url("../media/aboutCurve.svg");
+    background-image: url("../assets/aboutCurve.svg");
   }
 
   #app .lg\:bg-landing-page{
-    background-image: url("../media/background_landing_page.svg");
+    background-image: url("../assets/background_landing_page.svg");
   }
 
   #app .lg\:from-transparent{
@@ -86052,11 +86052,11 @@ video {
   }
 
   #app .xl\:bg-about{
-    background-image: url("../media/aboutCurve.svg");
+    background-image: url("../assets/aboutCurve.svg");
   }
 
   #app .xl\:bg-landing-page{
-    background-image: url("../media/background_landing_page.svg");
+    background-image: url("../assets/background_landing_page.svg");
   }
 
   #app .xl\:from-transparent{
