+++ src/assets/sandbox_square_salmon.png
