+++ src/landingPage/LandingPage.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import background_img from '../media/background.jpeg';
-import landing_background from '../media/background_landing_page.svg';
+import background_img from '../assets/background.jpeg';
+import landing_background from '../assets/background_landing_page.svg';
 import descriptions from '../media/modules/module_descriptions.json';
 import ModuleIntro from "./ModuleIntro";
