+++ src/modules/computerVision/common/KernelDisplay.tsx
